+++ config/webpack.dev.conf.js
@@ -1,7 +1,7 @@
 const wd = process.cwd()
 
 const path = require('path')
-const merge = require('webpack-merge')
+const { merge } = require('webpack-merge')
 const HtmlWebpackPlugin = require('html-webpack-plugin')
 
 const webpackBaseConfig = require('./webpack.base.conf')
@@ -9,26 +9,24 @@ const tidoryConfig = require('../tidory.config')
 
 module.exports = async env => {
   return merge(await webpackBaseConfig(env), {
-    devtool: 'inline-source-map',
-    resolve: {
-      alias: {
-        vue: path.resolve(wd, 'node_modules/vue/dist/vue.js')
-      }
-    },
+    mode: 'development',
     devServer: {
       watchContentBase: true,
       index: tidoryConfig.path.index,
+      hot: true,
       port: env.development ? '8080' : '3000',
-      stats: 'errors-only'
+      stats: 'normal'
     },
     output: {
-      filename: '[name].[hash].js'
+      filename: '[name].[fullhash].js'
     },
+    devtool: 'source-map',
     plugins: [
       new HtmlWebpackPlugin({
         template: path.join(wd, tidoryConfig.path.template),
         filename: tidoryConfig.path.index,
-        inject: true
+        inject: true,
+        scriptLoading: 'blocking'
       })
     ]
   })